+++ src/common/js/index.js
@@ -4,6 +4,7 @@
 
 import { DisplaySet } from '../classes/display-set.class.js';
 import { DisplayUnit } from '../classes/display-unit.class.js';
+import { MenuItem } from '../classes/menu-item.class.js';
 import { RunContext } from '../classes/run-context.class.js';
 
 import './global.js';
@@ -14,4 +15,5 @@ import './trace.js';
 // provides base classes in AppPages global object
 AppPages.DisplaySet = DisplaySet;
 AppPages.DisplayUnit = DisplayUnit;
+AppPages.MenuItem = MenuItem;
 AppPages.RunContext = RunContext;